+++ HISTORY.rst
@@ -7,3 +7,10 @@ History
 ++++++++++++++++++
 
 * First release on PyPI.
+* Basic Sponsor model
+* First version of templatetag
+* Some tests on model, view and templatetag
+* Optional default logo's width and height in settings
+* Basic styles
+* Admin customized
+* Includes a ListView of Sponsors